updating the model

+++ RepairShop-Backend/src/main/java/ca/mcgill/ecse321/repairshop/service/AppointmentService.java
@@ -1,6 +1,8 @@
 package ca.mcgill.ecse321.repairshop.service;
 
 import ca.mcgill.ecse321.repairshop.dao.AppointmentRepository;
+import ca.mcgill.ecse321.repairshop.dao.CustomerRepository;
+import ca.mcgill.ecse321.repairshop.dao.ServiceRepository;
 import ca.mcgill.ecse321.repairshop.dao.TimeSlotRepository;
 import ca.mcgill.ecse321.repairshop.model.*;
 import ca.mcgill.ecse321.repairshop.utility.RepairShopUtil;
@@ -17,9 +19,27 @@ public class AppointmentService {
     AppointmentRepository appointmentRepository;
     @Autowired
     TimeSlotRepository timeSlotRepository;
+    @Autowired
+    CustomerRepository customerRepository;
+    @Autowired
+    ServiceRepository serviceRepository;
 
     @Transactional
     public Appointment createAppointment(List<BookableService> services, Customer customer, TimeSlot timeslot) {
+        // IF EITHER ARGUMENT IS NULL
+        if (services == null || customer == null || timeslot == null){
+            throw new IllegalArgumentException("Customer, services and timeslot must all be selected for the appointment!");
+        }
+
+        // IF EITHER ARGUMENT IS NOT FOUND IN DATABASE
+        if (customerRepository.findCustomerById(customer.getId()) == null ||timeSlotRepository.findTimeSlotById(timeslot.getId()) == null){
+            throw new IllegalArgumentException("Bookable Service, Customer, Timeslot don't exist!");
+        }
+        for (BookableService s : services){
+            if (serviceRepository.findServiceById(s.getId())== null){
+                throw new IllegalArgumentException("Bookable Service, Customer, Timeslot don't exist!");
+            }
+        }
         Appointment appointment = new Appointment();
         appointment.setServices(services);
         appointment.setCustomer(customer);
@@ -34,6 +54,7 @@ public class AppointmentService {
         return appointment;
     }
 
+    // TODO: is this really needed?
     @Transactional
     public List<Appointment> getAllAppointment() {
         return RepairShopUtil.toList(appointmentRepository.findAll());
@@ -82,44 +103,20 @@ public class AppointmentService {
 
     @Transactional
     public void deleteAppointment (Appointment appointment){
-
+        if (appointment == null) {
+            throw new IllegalArgumentException("Cannot delete a null appointment");
+        }
         //delete appointment from the service
         appointmentRepository.deleteById(appointment.getId());
         // todo: need to delete from the customer, service as well?
     }
+
     @Transactional
     public List<Appointment> getAppointmentsBookedByCustomer(Customer customer) {
-        List<Appointment> appointmentsBookedByCustomer = new ArrayList<>();
-        for (Appointment app : appointmentRepository.findByCustomer(customer)) {
-            appointmentsBookedByCustomer.add(app);
-        }
+
+        List<Appointment> appointmentsBookedByCustomer = appointmentRepository.findByCustomer(customer);
         return appointmentsBookedByCustomer;
     }
 
-    @Transactional
-    public List<Appointment> getAppointmentsOfService(BookableService service_obj) {
-        List<Appointment> appointmentsOfService = new ArrayList<>();
-        for (Appointment app : appointmentRepository.findByServices(service_obj)) {
-            appointmentsOfService.add(app);
-        }
-        return appointmentsOfService;
-    }
 
-    @Transactional
-    public List<Appointment> getAppointmentsOfBill(Bill bill){
-        List<Appointment> appointmentsOfBill = new ArrayList<>();
-        for (Appointment app : appointmentRepository.findByBill(bill)) {
-            appointmentsOfBill.add(app);
-        }
-        return appointmentsOfBill;
-    }
-
-//    @Transactional
-//    public List<Appointment> getAppointmentsByTimeSlot (TimeSlot timeslot){
-//        List<Appointment> appointmentsOfTime = new ArrayList<>();
-//        for (Appointment app : appointmentRepository.findByTimeSlot(timeslot)) {
-//            appointmentsOfTime.add(app);
-//        }
-//        return appointmentsOfTime;
-//    }
 }
